Added exception to not remove index.html extension
Nginx was not being able to to display the base url showing index without its extension

+++ RemoveHtmlExtensionFromOutputFiles.py
@@ -4,11 +4,8 @@ from os import listdir
 import datetime
 
 """ Instructions
-Run the file with no arguments to convert the post note with today's date attribute
-OR
-Give an argument (date in format YYYY-mm-dd) to convert the note with that date attribute
-OR
-Give the argument "all" and it runs on all posts present (useful for some kind of global reformatting)
+Run the file without arguments
+Removes extension from all html output files
 """
 
 
@@ -21,5 +18,5 @@ output_files_directory = program_directory + "\\docs\\"
 
 # list_files = [f for f in listdir(vault_thoughts_directory) if (os.path.isfile(os.path.join(vault_thoughts_directory, f)) and f[-2:]) == "md"]
 for f in listdir(output_files_directory):
-    if (os.path.isfile(os.path.join(output_files_directory, f)) and f[-5:]) == ".html":
+    if (os.path.isfile(os.path.join(output_files_directory, f)) and f[-5:] == ".html" and f != "index.html"): # Excluding index.html because I couldn't get Nginx to see it without the extension
         os.rename(os.path.join(output_files_directory, f), os.path.join(output_files_directory, f[:-5]))
